requirements.txt
aws-cdk-lib==2.181.1
constructs>=10.0.0,<11.0.0
cdk_nag
bandit
pytest
pip-audit
requests==2.32.3